Separate exceptions into their own file
This avoids a circular dependency between __init__.py and the decoder
modules.
