+++ audioread/__init__.py
@@ -15,18 +15,8 @@
 """Decode audio files."""
 
 from . import ffdec
+from .exceptions import DecodeError, NoBackendError
 from .version import version as __version__  # noqa
-
-class DecodeError(Exception):
-    """The base exception class for all decoding errors raised by this
-    package.
-    """
-
-
-class NoBackendError(DecodeError):
-    """The file could not be decoded by any backend. Either no backends
-    are available or each available backend failed to decode the file.
-    """
 
 
 def _gst_available():
